refactor: create new test structure
