+++ src/test/java/ru/yandex/practicum/filmorate/ControllerTest.java
@@ -1,4 +1,28 @@
 package ru.yandex.practicum.filmorate;
 
-public abstract class ControllerTest {
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
+
+import javax.validation.Validation;
+import javax.validation.Validator;
+import javax.validation.ValidatorFactory;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+abstract class ControllerTest {
+    @Autowired
+    protected MockMvc mockMvc;
+    @Autowired
+    protected ObjectMapper objectMapper;
+    protected ValidatorFactory factory;
+    protected Validator validator;
+    @BeforeEach
+    void init() {
+        factory = Validation.buildDefaultValidatorFactory();
+        validator = factory.getValidator();
+    }
 }
